pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "causaganha"
version = "1.0.0"
description = "Automated judicial decision analysis using TrueSkill rating system"
authors = [{name = "CausaGanha Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.0",
    "duckdb>=0.10.0",
    "google-api-python-client>=2.173.0",
    "google-auth>=2.40.3",
    "google-generativeai>=0.8.5",
    "internetarchive>=5.4.0",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
    "python-dotenv>=1.0.0",
    "python-json-logger>=3.3.0",
    "requests>=2.32.4",
    "toml>=0.10.2",
    "trueskill>=0.4.5",
    "zstandard>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=4.0.0",
    "ruff>=0.12.0",
]

[project.scripts]
causaganha = "cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml"

[dependency-groups]
dev = [
    "pytest-cov>=6.2.1",
]
